refactor(cartridge): extract into self-contained looplet.cartridge package
Convert the cartridge code from sibling modules under src/looplet/
into a self-contained looplet.cartridge package whose dependency
direction is one-way: cartridge depends only on looplet's *public
types*, never on runtime internals. This is the prerequisite for
extracting looplet-cartridge into its own pip-installable package
(see paper/principled_cartridge_v2.md and architectural_assessment.md
for the rationale).

## Layout

  src/looplet/
  ├── cartridge/                 # NEW — self-contained package
  │   ├── __init__.py            # was src/looplet/cartridge.py (3313 lines)
  │   ├── scaffold.py            # was src/looplet/scaffold.py
  │   ├── spec_slots.py          # was src/looplet/spec_slots.py
  │   ├── prompt_files.py        # was src/looplet/prompt_files.py
  │   └── hot_reload.py          # was src/looplet/hot_reload.py
  ├── scaffold.py                # back-compat shim re-exporting from cartridge.scaffold
  ├── spec_slots.py              # back-compat shim
  ├── prompt_files.py            # back-compat shim
  └── hot_reload.py              # back-compat shim

All four shim modules at the historical paths re-export their
canonical objects from looplet.cartridge.* — existing user code
that imports from looplet.scaffold import scaffold_cartridge
keeps working unchanged.

## Dependency direction (what the new tests pin)

The cartridge package now depends only on the following public
looplet modules: builtin_hooks, builtin_tools, compact, hook_decision,
loop, memory, permissions, presets, refs, tools, types, validation.

Forbidden dependencies (caught by tests/test_cartridge_package_layout.py):
async_loop, context, recovery*, scaffolding (the LLM-call retry
helpers, distinct from scaffold), budget, streaming, telemetry,
evals, provenance, session, checkpoint, cache, router, subagent,
skills, bundles, rpc, parse, history, conversation.

## Why

* The paper claims any conformant runtime can load any cartridge.
  That claim is structurally defensible only if the cartridge code
  doesn't reach into one runtime's internals.
* The cartridge.py module was 3313 lines tangled with sibling
  modules — exactly the kind of growth the paper warns against
  ("the runtime contract is small enough to state on one page").
* Extracting into a package is the precondition for the future
  looplet-cartridge pip package, where looplet-cartridge has no
  hard dep on looplet itself (only on a small protocol surface).

## Other small repointings

* src/looplet/__init__.py and builtin_tools/scaffold_cartridge.py
  now import from the canonical looplet.cartridge.scaffold path
  (not via the back-compat shim), so the shim isn't load-bearing
  for our own code.
* The cartridge package's prompt_files imports from
  looplet.hook_decision directly (not via the looplet umbrella),
  avoiding a top-level circular dep.

## Tests (+6)

tests/test_cartridge_package_layout.py:
* canonical names exported by looplet.cartridge
* four submodules (scaffold, spec_slots, prompt_files, hot_reload)
  importable from canonical paths
* back-compat shims still work; same objects (is-equal)
* every external looplet dep on the public-type allowlist
* no deps on runtime internals (forbidden-deps test)
* no top-level from looplet import X inside the cartridge
  package (avoid umbrella circularity)

## Verification

* 1930 tests pass (was 1924; +6 new).
* ruff + format + pyright clean.
* git mv preserves history for all 5 moved files.
* cartridge package's external looplet deps: 12 (all on public types,
  none on internal machinery).
